refactor: split exporter.go into small files.
- export_processor.go
  Core logic for processing and exporting files from Synology Drive.
- exporter_core.go
  Main Exporter struct and constructor functions with dependency injection.
- file_operations.go
  File system operations including file removal and cleanup.
- filesystem.go
  Filesystem operation interfaces and default implementations.
- stats.go
  Statistics collection and reporting for export operations
